# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "ORM Capital Calculator Engine",
    "description": "RBI Basel III SMA Compliance System\n\nA comprehensive system for calculating operational risk capital requirements in compliance with RBI's Basel III Standardized Measurement Approach (SMA).\n\n## Features\n\n* **Job-based API**: Synchronous and asynchronous calculation execution\n* **Multiple Methods**: SMA, BIA, TSA, and what-if scenarios\n* **Data Lineage**: Complete audit trail and reproducibility\n* **Parameter Management**: Governed parameter updates with maker-checker workflow\n* **Webhook Notifications**: Real-time job completion notifications\n* **Comprehensive Validation**: Input validation with detailed error messages\n\n## Authentication\n\nThis API uses OAuth 2.0 client credentials flow for authentication.\nInclude the `Authorization: Bearer <token>` header in your requests.\n\n## Rate Limiting\n\nAPI requests are rate-limited to ensure system stability:\n* 100 requests per minute for calculation endpoints\n* 1000 requests per minute for status/health endpoints\n\n## Error Handling\n\nAll errors follow a standardized format:\n```json\n{\n    \"error_code\": \"ERROR_CODE\",\n    \"error_message\": \"Human readable message\",\n    \"details\": {\n        \"additional\": \"context\"\n    }\n}\n```",
    "version": "0.1.0",
    "contact": {
      "name": "ORM Capital Calculator Support",
      "email": "support@example.com"
    },
    "license": {
...
    }
...
  "servers": [
...
  "paths": {
...
    "/calculation-jobs": {
      "post": {
        "tags": ["calculations"],
        "summary": "Create Calculation Job",
        "description": "Create a new calculation job (sync or async)",
        "operationId": "create_calculation_job",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Synchronous calculation completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationJobResponse"
                }
              }
            }
          },
          "202": {
            "description": "Asynchronous job queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calculation-jobs/{job_id}": {
      "get": {
        "tags": ["calculations"],
        "summary": "Get Job Status",
        "description": "Retrieve job status and results",
        "operationId": "get_job_status",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status and results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CalculationRequest": {
        "type": "object",
        "properties": {
          "model_name": {
            "type": "string",
            "enum": ["sma", "bia", "tsa", "what-if"]
          },
          "execution_mode": {
            "type": "string",
            "enum": ["sync", "async"]
          },
          "entity_id": {
            "type": "string"
          },
          "calculation_date": {
            "type": "string",
            "format": "date"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": true
          },
          "callback_url": {
            "type": "string",
            "format": "uri"
          },
          "idempotency_key": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string"
          }
        },
        "required": ["model_name", "execution_mode", "entity_id", "calculation_date"]
      },
      "CalculationJobResponse": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["completed", "failed"]
          },
          "results": {
            "$ref": "#/components/schemas/CalculationResults"
          },
          "run_id": {
            "type": "string"
          },
          "calculation_date": {
            "type": "string",
            "format": "date"
          },
          "entity_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "processing_time_seconds": {
            "type": "number"
          }
        },
        "required": ["job_id", "status", "run_id"]
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "running"]
          },
          "estimated_completion_time": {
            "type": "string",
            "format": "date-time"
          },
          "callback_url": {
            "type": "string",
            "format": "uri"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["job_id", "status", "created_at"]
      },
      "JobStatusResponse": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "completed", "failed"]
          },
          "progress_percentage": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "processing_time_seconds": {
            "type": "number"
          },
          "result_available": {
            "type": "boolean"
          },
          "results": {
            "$ref": "#/components/schemas/CalculationResults"
          },
          "run_id": {
            "type": "string"
          },
          "error_details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["job_id", "status", "result_available"]
      },
      "CalculationResults": {
        "type": "object",
        "properties": {
          "business_indicator": {
            "type": "number",
            "description": "Business Indicator (BI) amount"
          },
          "business_indicator_component": {
            "type": "number",
            "description": "Business Indicator Component (BIC) amount"
          },
          "loss_component": {
            "type": "number",
            "description": "Loss Component (LC) amount"
          },
          "internal_loss_multiplier": {
            "type": "number",
            "description": "Internal Loss Multiplier (ILM)"
          },
          "operational_risk_capital": {
            "type": "number",
            "description": "Operational Risk Capital (ORC) amount"
          },
          "risk_weighted_assets": {
            "type": "number",
            "description": "Risk Weighted Assets (RWA) amount"
          },
          "bucket": {
            "type": "integer",
            "enum": [1, 2, 3],
            "description": "SMA bucket classification"
          },
          "methodology": {
            "type": "string",
            "enum": ["sma", "bia", "tsa"],
            "description": "Calculation methodology used"
          },
          "calculation_metadata": {
            "type": "object",
            "properties": {
              "ilm_gate_applied": {
                "type": "boolean"
              },
              "data_quality_score": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              "loss_events_included": {
                "type": "integer"
              }
            }
          }
        },
        "required": ["operational_risk_capital", "risk_weighted_assets", "methodology"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error_code": {
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["error_code", "error_message"]
      }
    }
  }
}